added time provider code and fixed clock_tick() issues
git-svn-id: http://sbfsvn02/gemini-sw/gem/trunk/support/bancomm@607 303ae0b8-5cfe-405b-ba84-a597b98f2cc0

+++ bancommApp/src/bc635.h
@@ -64,8 +64,8 @@ extern int	bcSendTfp(char *);
 extern void	bcSetRTC(void);
 extern int	bcSetEpoch (const int);
 extern int	bcRegsToTime (double *, unsigned char *);
-extern int	bc635_report (int);
-extern int	bc635_init (void);
+extern long	bc635_report (int);
+extern long	bc635_init (void);
 extern int      bcTestCard( void );
 extern int      bcGetGpsLeap( void );
